+++ catchBullying/Models/PsicologistQuestion.swift
@@ -1,0 +1,13 @@
+//
+//  PsicologistQuestions.swift
+//  catchBullying
+//
+//  Created by apple on 21/05/1443 AH.
+//
+
+import Foundation
+struct PsicologistQuestion {
+let question: String
+let answers: [String]
+  
+}
