--- catchBullying/Models/PsicologistQuestions.swift
@@ -1,13 +1,0 @@
-//
-//  PsicologistQuestions.swift
-//  catchBullying
-//
-//  Created by apple on 21/05/1443 AH.
-//
-
-import Foundation
-struct PsicologistQuestions {
-let question: String
-let answers: [String]
-  
-}
