+++ catchBullying/View/TableViewCell.swift
@@ -1,0 +1,23 @@
+//
+//  TableViewCell.swift
+//  catchBullying
+//
+//  Created by apple on 08/06/1443 AH.
+//
+
+import UIKit
+
+class TableViewCell: UITableViewCell {
+
+    override func awakeFromNib() {
+        super.awakeFromNib()
+        // Initialization code
+    }
+
+    override func setSelected(_ selected: Bool, animated: Bool) {
+        super.setSelected(selected, animated: animated)
+
+        // Configure the view for the selected state
+    }
+    
+}
