+++ catchBullying/ViewControllers/Chat/ChatTableViewController.swift
@@ -16,6 +16,8 @@ class ChatTableViewController: UITableViewController {
   
   override func viewDidLoad() {
     super.viewDidLoad()
+
+
   }
   
   override func viewWillAppear(_ animated: Bool) {
