+++ catchBullying/ViewControllers/Chat/MessageCell.swift
@@ -15,6 +15,8 @@ class MessageCell: UITableViewCell {
   
   override func awakeFromNib() {
         super.awakeFromNib()
+   
+
         // Initialization code
     }
