+++ catchBullying/ViewControllers/Doctor/DoctorInformtionViewController.swift
@@ -47,7 +47,8 @@ class DoctorInformtionViewController: UIViewController {
   
   override func viewDidLoad() {
     super.viewDidLoad()
-    
+   
+
