+++ catchBullying/ViewControllers/Patient/DoctorsListCell.swift
@@ -17,6 +17,8 @@ class DoctorsListCell: UITableViewCell {
   override func awakeFromNib() {
         super.awakeFromNib()
         // Initialization code
+   
+
     }
 
     override func setSelected(_ selected: Bool, animated: Bool) {
