+++ catchBullying/ViewControllers/Patient/DoctorsListViewController.swift
@@ -21,7 +21,8 @@ class DoctorsListViewController: UIViewController {
   
   override func viewDidLoad() {
     super.viewDidLoad()
-    
+   
+
     tableView.delegate = self
     tableView.dataSource = self
     
@@ -78,7 +79,7 @@ extension DoctorsListViewController : UITableViewDelegate , UITableViewDataSourc
     //   cell.textLabel?.text = arryListOfDoctores[indexPath.row].zoom
     //cell.imageView?.image = data[indexPath.row].image
     //cell.textLabel?.text = data[indexPath.row].firstName
-    //cell.imageViewDoctor.layer.cornerRadius = cell.imageViewDoctor.frame.width/2
+    cell.imageViewDoctor.layer.cornerRadius = cell.imageViewDoctor.frame.width/2
     
     return cell
   }
