+++ catchBullying/ViewControllers/Patient/PatientInformtionViewController.swift
@@ -30,7 +30,8 @@ class PatientInformtionViewController: UIViewController {
   
   override func viewDidLoad() {
     super.viewDidLoad()
-    
+ 
+
     tableView.delegate = self
     tableView.dataSource = self
     tableView.register(InformationCell.self, forCellReuseIdentifier: "cell")
