+++ catchBullying/ViewControllers/Patient/ProfileViewController.swift
@@ -54,6 +54,8 @@ class ProfileViewController: UIViewController {
   
   override func viewDidLoad() {
     super.viewDidLoad()
+ 
+
     tableView.delegate = self
     tableView.dataSource = self
     tableView.register(UITableViewCell.self, forCellReuseIdentifier: "cell")
@@ -148,7 +150,7 @@ extension ProfileViewController: UITableViewDelegate, UITableViewDataSource {
   func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
     let cell = tableView.dequeueReusableCell(withIdentifier: "cell", for: indexPath)
     cell.imageView?.image = UIImage(systemName: data[indexPath.section][indexPath.row].icon)
-    cell.textLabel?.text = data[indexPath.section][indexPath.row].title
+    cell.textLabel?.text = NSLocalizedString(data[indexPath.section][indexPath.row].title, comment: "")
     cell.textLabel?.textColor = data[indexPath.section][indexPath.row].color
     return cell
   }
