+++ catchBullying/ViewControllers/Patient/QuestionsViewController.swift
@@ -6,19 +6,28 @@
 //
 
 import UIKit
+import Firebase
+import FirebaseFirestoreSwift
 
 class QuestionsViewController: UIViewController {
-  var currentQuestionIndex: Int = 0
-  var answers: [String] = []
+  var current: Int = 0
+  var answers: [Int] = []
   
-  let arryPsicologistQuestions : [PsicologistQuestions] = [
-    PsicologistQuestions(question: "Gender", answers:["Male" , "Female" ,"Other"]),
-    PsicologistQuestions(question: "Do you currentiy have a job?", answers:["yes" , "No" ,"prefer not answer"]),
-    PsicologistQuestions(question: "Have you taken a psychiatric medication befor?", answers: ["yes" , "No" ,"prefer not answer"]),
-    PsicologistQuestions(question: "Does your family have a history of psychiatrist counseling?", answers: ["yes" , "No" ,"prefer not answer"]),
-    PsicologistQuestions(question: "Were you subjected to domestic violence (verbal , psychological, physical, sexual , negligence ...)?", answers: ["yes" , "No" ,"prefer not answer"]),
-    PsicologistQuestions(question: "Do you think of hurting yourself or have suicidal thoughts ?", answers: ["yes" , "No" ,"prefer not answer"]),
-    PsicologistQuestions(question: "Do you plan on committing suicide ??", answers: ["yes" , "No" ,"prefer not answer"])
+  let questions: [PsicologistQuestion] = [
+    PsicologistQuestion(question: "Gender",
+                        answers:["Male" , "Female" ,"Other"]),
+    PsicologistQuestion(question: "Do you currentiy have a job?",
+                        answers:["Yes" , "No" ,"Prefer not answer"]),
+    PsicologistQuestion(question: "Have you taken a psychiatric medication befor?",
+                        answers: ["Yes" , "No" ,"Prefer not answer"]),
+    PsicologistQuestion(question: "Does your family have a history of psychiatrist counseling?",
+                        answers: ["Yes" , "No" ,"Prefer not answer"]),
+    PsicologistQuestion(question: "Were you subjected to domestic violence (verbal , psychological, physical, sexual , negligence ...)?",
+                        answers: ["Yes" , "No" ,"Prefer not answer"]),
+    PsicologistQuestion(question: "Do you think of hurting yourself or have suicidal thoughts?",
+                        answers: ["Yes" , "No" ,"Prefer not answer"]),
+    PsicologistQuestion(question: "Do you plan on committing suicide?",
+                        answers: ["Yes" , "No" ,"Prefer not answer"])
     
   ]
   
@@ -29,38 +38,49 @@ class QuestionsViewController: UIViewController {
   @IBOutlet weak var nextButton: UIButton!
   
   override func viewDidLoad() {
+    
     super.viewDidLoad()
     self.isModalInPresentation = true
     self.navigationItem.setHidesBackButton(true, animated: false)
     configureButtons()
-    questionLabel.text=arryPsicologistQuestions[currentQuestionIndex].question
+    setupQuestion()
   }
   
   func setupQuestion() {
-    questionLabel.text = arryPsicologistQuestions[currentQuestionIndex].question
+    questionLabel.text = questions[current].question
+    yesButton.setTitle(questions[current].answers[0], for: .normal)
+    noButton.setTitle(questions[current].answers[1], for: .normal)
+    otherButton.setTitle(questions[current].answers[2], for: .normal)
   }
   
   func nextQuestion() {
-    print(answers)
-    if currentQuestionIndex < arryPsicologistQuestions.count - 1 {
-      currentQuestionIndex += 1
+    if current < questions.count - 1 {
+      current += 1
       setupQuestion()
     } else {
-      print("dismiss")
       if isUpdating {
         self.dismiss(animated: true, completion: nil)
       } else {
-        if user.isDoctor {
-          let controller = self.storyboard?.instantiateViewController(identifier: "DoctorHomeVC") as! DoctorHomeViewController
-          controller.modalPresentationStyle = .fullScreen
-          controller.modalTransitionStyle = .flipHorizontal
-          self.present(controller, animated: false, completion: nil)
-        } else {
-          let controller = self.storyboard?.instantiateViewController(identifier: "UserHomeVC") as! UserHomeViewController
-          controller.modalPresentationStyle = .fullScreen
-          controller.modalTransitionStyle = .flipHorizontal
-          self.present(controller, animated: false, completion: nil)
+        patientProfile.answers = answers
+        let db = Firestore.firestore()
+        do {
+          try db.collection("patients").document(patientProfile.docID!).setData(from: patientProfile,
+                                                                                merge: true) { error in
+                if let error = error {
+                  fatalError(error.localizedDescription)
+                }
+                
+                let controller = self.storyboard?.instantiateViewController(identifier: "UserHomeVC") as! UserHomeViewController
+                controller.modalPresentationStyle = .fullScreen
+                controller.modalTransitionStyle = .flipHorizontal
+                self.present(controller, animated: false, completion: nil)
+              }
+        } catch {
+          fatalError(error.localizedDescription)
         }
+        
+        
+        
       }
       
       
@@ -69,12 +89,13 @@ class QuestionsViewController: UIViewController {
   
   
   @IBAction func nextButton(_ sender: Any) {
+    answers.append(-1)
     nextQuestion()
   }
   
   
   @IBAction func answuer(_ sender: UIButton) {
-    answers.append(sender.titleLabel!.text!)
+    answers.append(sender.tag)
     nextQuestion()
   }
   
@@ -82,14 +103,17 @@ class QuestionsViewController: UIViewController {
   
   
   func configureButtons() {
+    yesButton.tag = 0
     yesButton.layer.cornerRadius = 20
     yesButton.layer.borderWidth = 3
     yesButton.layer.borderColor = CGColor(red: 225, green:255, blue: 255, alpha: 1)
     
+    noButton.tag = 1
     noButton.layer.cornerRadius = 20
     noButton.layer.borderWidth = 3
     noButton.layer.borderColor = CGColor(red: 225, green:255, blue: 255, alpha: 1)
     
+    otherButton.tag = 2
     otherButton.layer.cornerRadius = 20
     otherButton.layer.borderWidth = 3
     otherButton.layer.borderColor = CGColor(red: 225, green:255, blue: 255, alpha: 1)
