+++ catchBullying/ViewControllers/Patient/SupportViewController.swift
@@ -15,7 +15,7 @@ class SupportViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
       
-      overrideUserInterfaceStyle = .light
+    
     }
   
   override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
