+++ catchBullying/en.lproj/Localizable.strings
@@ -1,0 +1,7 @@
+/* 
+  Localizable.strings
+  catchBullying
+
+  Created by apple on 07/06/1443 AH.
+  
+*/
